Code tidy up and reorganisation of index.js. Added handlers folder to group routes logically
