+++ lib/domain.js
@@ -1,5 +1,5 @@
 //Adding middleware to create an app domain. This allows for catching uncaught exceptions
-// a domain is basically a an execution context that will catch errors which occur inside it
+// a domain is basically an execution context that will catch errors which occur inside it
 
 module.exports = function(app) {
