+++ lib/middleware.js
@@ -1,5 +1,5 @@
 //**- Middleware --**
-module.exports = function (app, express) {
+module.exports = function (app) {
     var bodyParser = require('body-parser');
 
     app.use(bodyParser.urlencoded({
