+++ models/sample.js
@@ -1,6 +1,6 @@
 // Require mongoose
 var mongoose = require('mongoose');
-var vacationSchema = mongoose.Schema( {
+var sampleSchema = mongoose.Schema( {
    name: String, // String
     number: Number, // Number
     tags: [String], // Array of strings 
@@ -8,6 +8,6 @@ var vacationSchema = mongoose.Schema( {
     
 });
 
-var Sample = mongoose.model('Sample', vacationSchema); // create a model using Mongoose - at this point Sample becomes like a class
+var Sample = mongoose.model('Sample', sampleSchema); // create a model using Mongoose - at this point Sample becomes like a class
 module.exports = Sample; // add the Sample object to exports this can then be imported like this
 // var Sample = require('./models/sample.js');
