+++ views/page.handlebars
@@ -1,0 +1,1 @@
+<h3>{{message}}</h3>
